# Back-end/swagger_server/models/like.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Like(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, image_id: str=None, user_id: str=None, time: datetime=None, status: str=None):  # noqa: E501
        """Like - a model defined in Swagger

        :param image_id: The image_id of this Like.  # noqa: E501
        :type image_id: str
        :param user_id: The user_id of this Like.  # noqa: E501
        :type user_id: str
        :param time: The time of this Like.  # noqa: E501
        :type time: datetime
        :param status: The status of this Like.  # noqa: E501
        :type status: str
        """
        self.swagger_types = {
            'image_id': str,
            'user_id': str,
            'time': datetime,
            'status': str
        }

        self.attribute_map = {
            'image_id': 'imageID',
            'user_id': 'userID',
            'time': 'time',
            'status': 'status'
        }

        self._image_id = image_id
        self._user_id = user_id
        self._time = time
        self._status = status

    @classmethod
    def from_dict(cls, dikt) -> 'Like':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Like of this Like.  # noqa: E501
        :rtype: Like
        """
        return util.deserialize_model(dikt, cls)

    @property
    def image_id(self) -> str:
        """Gets the image_id of this Like.


        :return: The image_id of this Like.
        :rtype: str
        """
        return self._image_id

    @image_id.setter
    def image_id(self, image_id: str):
        """Sets the image_id of this Like.


        :param image_id: The image_id of this Like.
        :type image_id: str
        """

        self._image_id = image_id

    @property
    def user_id(self) -> str:
        """Gets the user_id of this Like.


        :return: The user_id of this Like.
        :rtype: str
        """
        return self._user_id

    @user_id.setter
    def user_id(self, user_id: str):
        """Sets the user_id of this Like.


        :param user_id: The user_id of this Like.
        :type user_id: str
        """

        self._user_id = user_id

    @property
    def time(self) -> datetime:
        """Gets the time of this Like.


        :return: The time of this Like.
        :rtype: datetime
        """
        return self._time

    @time.setter
    def time(self, time: datetime):
        """Sets the time of this Like.


        :param time: The time of this Like.
        :type time: datetime
        """

        self._time = time

    @property
    def status(self) -> str:
        """Gets the status of this Like.


        :return: The status of this Like.
        :rtype: str
        """
        return self._status

    @status.setter
    def status(self, status: str):
        """Sets the status of this Like.


        :param status: The status of this Like.
        :type status: str
        """
        allowed_values = ["active", "inactive"]  # noqa: E501
        if status not in allowed_values:
            raise ValueError(
                "Invalid value for `status` ({0}), must be one of {1}"
                .format(status, allowed_values)
            )

        self._status = status