# Back-end/swagger_server/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "comp9900"
  version: "1.0.0"
  title: "comp9900"
host: "127.0.0.1"
basePath: "/v1"
...
schemes:
- "https"
- "http"
paths:
...
  /image/getImage:
    get:
      tags:
      - "image"
      summary: "return default display photo/search image using keyword"
      description: "N"
      operationId: "get_image_by_keyword"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/Search"
      responses:
        "200":
          description: "successfully get images"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Image"
        "404":
          description: "not found"
          schema:
            type: "string"
      x-swagger-router-controller: "swagger_server.controllers.image_controller"
...
  /image/toplike:
    get:
      tags:
      - "image"
      summary: "return toplike photo"
      description: "N"
      operationId: "toplike_photo"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "successfully get info"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Image"
        "404":
          description: "not found"
          schema:
            type: "string"
      x-swagger-router-controller: "swagger_server.controllers.image_controller"
  /image/topcomment:
    get:
      tags:
      - "image"
      summary: "return topcomment photo"
      description: "N"
      operationId: "topcomment_photo"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "successfully get info"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Image"
        "404":
          description: "not found"
          schema:
            type: "string"
      x-swagger-router-controller: "swagger_server.controllers.image_controller"
  /image/bestsell:
    get:
      tags:
      - "image"
      summary: "return bestsell photo"
      description: "N"
      operationId: "bestsell_photo"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "successfully get info"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Image"
        "404":
          description: "not found"
          schema:
            type: "string"
      x-swagger-router-controller: "swagger_server.controllers.image_controller"
  /image/detail/{imageID}:
    get:
      tags:
      - "image"
      summary: "return image detail"
      description: "N"
      operationId: "image_detail"
      produces:
      - "application/json"
      parameters:
      - name: "imageID"
        in: "path"
        description: "image ID"
        required: true
        type: "string"
      responses:
        "200":
          description: "get photo detail"
          schema:
            $ref: "#/definitions/Image"
        "404":
          description: "Not found"
          schema:
            type: "string"
      x-swagger-router-controller: "swagger_server.controllers.image_controller"
    put:
      tags:
      - "image"
      summary: "update image detail"
      description: "N"
      operationId: "image_update"
      produces:
      - "application/json"
      parameters:
      - name: "imageID"
        in: "path"
        description: "image ID"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/Image"
      responses:
        "200":
          description: "update photo detail"
          schema:
            type: "string"
        "404":
          description: "Not found"
          schema:
            type: "string"
      x-swagger-router-controller: "swagger_server.controllers.image_controller"
  /image:
    post:
      tags:
      - "image"
      summary: "upload image"
      description: "N"
      operationId: "image_upload"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/Image"
      responses:
        "200":
          description: "upload photo detail"
          schema:
            type: "string"
        "400":
          description: "Bad request"
          schema:
            type: "string"
      x-swagger-router-controller: "swagger_server.controllers.image_controller"
  /image/like:
    post:
      tags:
      - "image"
      summary: "like an image"
      description: "N"
      operationId: "image_like"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/Like"
      responses:
        "200":
          description: "like photo"
          schema:
            type: "string"
        "400":
          description: "Bad request"
          schema:
            type: "string"
      x-swagger-router-controller: "swagger_server.controllers.image_controller"
  /image/comment:
    post:
      tags:
      - "image"
      summary: "comment an image"
      description: "N"
      operationId: "image_comment"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/Comment"
      responses:
        "200":
          description: "comment photo"
          schema:
            type: "string"
        "400":
          description: "Bad request"
          schema:
            type: "string"
      x-swagger-router-controller: "swagger_server.controllers.image_controller"
...
definitions:
  Tag:
    type: "object"
    properties:
      id:
        type: "string"
      name:
        type: "string"
    example:
      name: "name"
      id: "id"
...
  Image:
    type: "object"
    properties:
      id:
        type: "string"
      title:
        type: "string"
      price:
        type: "integer"
      userID:
        type: "string"
      status:
        type: "string"
        enum:
        - "on shop"
        - "off shop"
      tags:
        type: "array"
        items:
          $ref: "#/definitions/Tag"
    example:
      price: 0
      id: "id"
      title: "title"
      userID: "userID"
      status: "on shop"
      tags:
      - name: "name"
        id: "id"
      - name: "name"
        id: "id"
...
  Like:
    type: "object"
    properties:
      imageID:
        type: "string"
      userID:
        type: "string"
      time:
        type: "string"
        format: "date-time"
      status:
        type: "string"
        enum:
        - "active"
        - "inactive"
    example:
      imageID: "imageID"
      time: "2000-01-23T04:56:07.000+00:00"
      userID: "userID"
      status: "active"
  Comment:
    type: "object"
    properties:
      imageID:
        type: "string"
      userID:
        type: "string"
      time:
        type: "string"
        format: "date-time"
      status:
        type: "string"
        enum:
        - "active"
        - "inactive"
      detail:
        type: "string"
    example:
      imageID: "imageID"
      time: "2000-01-23T04:56:07.000+00:00"
      detail: "detail"
      userID: "userID"
      status: "active"
...
  Search:
    type: "object"
    properties:
      userID:
        type: "string"
      keyword:
        type: "string"
      time:
        type: "string"
        format: "date-time"
      searchType:
        type: "string"
        enum:
        - "Image"
        - "Creater"
...